refactor: categories.js 파일명을 constants.js로 변경
